Three changes in there
1) Concerning the normalization of the transmission vector, it is now smoothed to find
a robust maximum as a normalization factor. The output vector itself is not smoothed.

2) calibration coefficient was (incorrectly) dividing the measured science spectrum, instead
of multiplying it. This has been fixed.
Indeed, it is defined as the simulated flux (computed as an integral over the theoretical spectrum) divided by
the measured flux in the standard images

3) Now computes the incidence angle at the position of the star in routine get_flux_calibration_coeff()
This used to be assumed at the center of the image. This has a small effect on the channel size, and hence on the
value of the integrated flux.