SourceExtractor class: several changes:
unsqueeze last dimension to avoid problems in case of reading just one image.
This can happen if the residual of the number of images modulo the number of cpus is 1.

Also changed the equation using modulation_ratio to compute the photometry of the modulated
and unmodulated parts (difference and sum) of the two cameras.
This is important as it directly impacts the absolute calibration of the standard star spectrum.
It makes it coherent with the equation in get_flux_calibration_coeff()