Two changes here:
Increased (temporarily ??) vignet box around each star, centered on location of first image,
so that internal cube alignment for each camera succeeds even for large offsets. This was
useful for one particular cube.

Second change adds a reasonable guess for the first order phase vector in the case
the "improved phase" vector (that contains orders >= 1) is not present.